vse: fix cert warnings
Using this patch we are fixing below cert warnings :-
1. cert_int31_c_violation: Casting
   g_crypto_to_ivc_map[aes_ctx->node_id].se_engine from
   unsigned int to unsigned char without checking its
   value may result in lost or misinterpreted data.

CID 587675
CID 587674
CID 587672
CID 587671
CID 587670
CID 587668

Bug 3512545

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2823956
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/crypto/tegra-hv-vse-safety.c
@@ -408,7 +408,7 @@ struct tegra_virtual_se_aes_req_context {
 	/* Operation type */
 	bool encrypt;
 	/* Engine id */
-	u8 engine_id;
+	uint32_t engine_id;
 };
 
 enum se_engine_id {
